requirements.txt
pymongo~=4.6.0
flask~=3.0.0
python-dotenv~=1.0.0
kubernetes~=27.2.0